Add basic controls
The control prototype is actually done. By moving the mouse you control a
cursor that moves around the "arena". By clicking and holding you "shoot"
a wall out from the cursor, and by dragging the mouse you can bend that
line in any direction as desired. Releasing will eventually build the
wall, but that remains to be implemented.